pyproject.toml
[project]
name = "hyperquant"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "hyperliquid-python-sdk>=0.4",
    "pandas>=2.0",
    "numpy>=1.24",
    "ta>=0.11",
    "aiosqlite>=0.19",
    "python-telegram-bot>=21.0",
    "structlog>=24.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
    "ruff>=0.8",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "UP",   # pyupgrade
    "B",    # flake8-bugbear
    "SIM",  # flake8-simplify
    "S",    # flake8-bandit (security)
]
ignore = [
    "S101",  # assert usage (fine in tests)
    "S104",  # binding to all interfaces
    "S311",  # pseudo-random generators (not used for crypto)
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S"]
